ml-model/requirements.txt
fastapi
uvicorn
python-multipart
ultralytics
Pillow
transformers
